fix(atomic): fix search disabled check for query suggestion (#3669)
* In atomic-search-box, remove `@State() private isSearchDisabled` and
instead add a `isSearchDisabledForEndUser` function which returns true
if either the `disableSearch` prop is set to true, or if the content of
the input is not long enough.
* Modify some tests in cypress regarding this.

The reason to have a function to check this is because it's possible
that `disableSearch` prop is controlled by implementer for other reasons
than the content of the input.

So it makes sense to have the "disabled because of the generic prop" to
be decoupled from the "disabled because of the length of the input", and
not conflate them with one another.

Hopefully with the name of the function, it makes it clearer why/when
the input actually get disabled for the end user, and that using the
prop directly to derive the state of the input, or control if query
suggestions should be requested or not is clearer.


https://coveord.atlassian.net/browse/KIT-3020