feat(headless/commerce): set currency type as currency code iso4217 instead of string (#3667)
Jira: https://coveord.atlassian.net/browse/LENS-1589

This branch is to change the type of the currency set in the commerce
state (context-state). It was a `string`, it's now switch to
`CurrencyCodeISO4217`.

By doing this, the value set by the user for the commerce context will
be type-checked differently (illustrate by the demo below):


https://github.com/coveo/ui-kit/assets/58052881/472fbfd2-b1c6-415e-9107-2076b997d592