feat(headless): executeSearch action for insight refactored to be compatible with the new event protocol (#3678)
## [SFINT-5418](https://coveord.atlassian.net/browse/SFINT-5418)

In this PR:

- I adapted the `ExecuteSearch` action for the insight use case to be
compatible with the new event protocol, instead of accepting an insight
analytics action as a parameter now it accepts a transitive action:
  ```
  interface TransitiveInsightSearchAction {
    legacy: LegacyInsightAction;
    next?: InsightSearchAction;
  }
  ```
when the analytics mode will be set to `legacy`, the `ExecuteSearch`
action will keep working exactly how it was before by executing the
search and logging a search event.
However, when the analytics mode will be set to `next`, we will no
longer send search event.

- I refactored the logic executed inside the `ExecuteSearch` insight
action to make it follow the pattern of the `ExecuteSearch` search
action.
This new pattern implements the class
`AsyncInsightSearchThunkProcessor`, this class will take care of doing
the search query and will take care of properly processing its response.
This new logic allows us to better structure and organize the code and
it makes it easier to test the logic executed by the `ExecuteSearch`
insight action

[SFINT-5418]:
https://coveord.atlassian.net/browse/SFINT-5418?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Sylvie Allain <58052881+sallainCoveo@users.noreply.github.com>