fix(atomic): fix query suggestions mismatch on searchbox disabled + manual input clear (#3699)
As described in the JIRA, the previous fix was partial: When the end
user clears the input "manually" (by erasing the content with backspace)
to go back under `minimumQueryLength`, then types again, there is a
query suggestion mismatch that appears briefly.

Proposed fix is to always `clearSuggestions()` on input if search box is
disabled.

https://coveord.atlassian.net/browse/KIT-3020