feat(headless/commerce): allow purchasing from cart controller (#3654)
Jira: https://coveord.atlassian.net/browse/LENS-1498

This branch is to add an action to log the `ec.purchase` when the
`purchase` function from the CartController is called.

Demo on Barca sports (linked to this local build)


https://github.com/coveo/ui-kit/assets/58052881/9f713808-f7d0-4a6c-b16f-c117fb06c411

Open question: it would be consistent to emit a `ec.purchase` event wih
an empty array of product (you basically buy nothing), it would be a
valid event. Do we have some situations that it does make sense, or
should we update the schema?