test(quantic): bare bones to start testing the new event protocol in quantic added (#3696)
## [SFINT-5424](https://coveord.atlassian.net/browse/SFINT-5424)

in this PR:

- Added the logic that will intercept analytics sent using the event
protocol.
- Created the `next-analytics-expectations.ts`, this file will be a home
to all the next analytics expectations, created the first expectation as
an example.
- Created `analytics-utils.ts` which contain the utility that allow us
to enable sending the analytics according to the event protocol.

This is the base to all the PRs that will come after where we will
update our test to check the analytics with the new event protocol.

[SFINT-5424]:
https://coveord.atlassian.net/browse/SFINT-5424?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Sylvie Allain <58052881+sallainCoveo@users.noreply.github.com>