feat(headless): add isAnswerGenerated to the state (#3740)
[SVCC-3597](https://coveord.atlassian.net/browse/SVCC-3597)

Add `isAnswerGenerated` to GenerateAnswerState, to be able to check if
the answer was properly generated.

`isAnswerGenerated` is set to `true` when an answer is properly
generated.
`isAnswerGenerated` is set to `false` when “sorry message” is returned.

<img width="946" alt="Screenshot 2024-03-21 at 8 22 20 PM"
src="https://github.com/coveo/ui-kit/assets/119955059/f14f95ae-f5da-4fcb-950e-2932d52c2718">


[SVCC-3597]:
https://coveord.atlassian.net/browse/SVCC-3597?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ