feat(quantic): disclaimer message for rga (#3739)
Added a disclaimer at the bottom of the RGA message.

The disclaimer has a default look and text of : 

`Generated content may contain errors. Verify important information.`

And the section for the disclaimer is also a slot, so if someone wants
to completely change what gets displayed as the disclaimer, they can not
only edit the text, but also the HTML.

The message is also a label, so it can be translated, etc.

Single line footer:

![image](https://github.com/coveo/ui-kit/assets/2488155/116c546a-0bdf-4e97-8375-73911a409efa)

Multi-line footer:

![image](https://github.com/coveo/ui-kit/assets/2488155/9ea7eb59-6bd1-4dea-9301-ef1ae14cb609)

In the Insight Panel:

![image](https://github.com/coveo/ui-kit/assets/2488155/e09e6b6b-cbe5-4a19-bc90-74266a629044)