fix(quantic): avoid pushing to recent results when use case is not search (#3731)
### [SFINT-5456](https://coveord.atlassian.net/browse/SFINT-5456)

The recent results feature is not yet supported for the insight use
case, I consider this as a temporary fix.
A real improvement to improve this would be the following:

- Make the necessary updates in the Headless library to support this
feature in the insight use case, we should also make the feature about
the recent queries list available for the insight use case.
- Improve the push recent action in headless v3, for now it takes the
whole result as a parameter, while just sending the unique id should be
enough to find the result in the Headless state.

#### BEFORE


https://github.com/coveo/ui-kit/assets/86681870/1d9dbd5f-ba2e-496f-b1cb-caee4d416708

#### AFTER


https://github.com/coveo/ui-kit/assets/86681870/cdea0d77-5d92-4739-85e1-9ba396b7e0ce



[SFINT-5456]:
https://coveord.atlassian.net/browse/SFINT-5456?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Sylvie Allain <58052881+sallainCoveo@users.noreply.github.com>
Co-authored-by: Etienne Rocheleau <erocheleau@coveo.com>