test(quantic): tests added to check for the analytics section in the search queries made (#3755)
## [SFINT-5422](https://coveord.atlassian.net/browse/SFINT-5422)

In this PR:

- I remade the tests that runs on the example search page, improved them
and added the tests to check for the analytics section sent in the
search request.
- I created new tests that runs on the example insight panel and added
tests that check the presence of the analytics section sent in the
insight search request.
- I added some additional checks in some components to confirm the
presence of the analytics section sent in the search requests for the
insight use case and the search use case.

Tests are about this section: 
<img width="724" alt="Screenshot 2024-03-28 at 8 45 12 AM"
src="https://github.com/coveo/ui-kit/assets/86681870/16e327d6-7be4-439b-89be-1cc418aaf6c8">


[SFINT-5422]:
https://coveord.atlassian.net/browse/SFINT-5422?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Sylvie Allain <58052881+sallainCoveo@users.noreply.github.com>