chore(headless): applying abortable fetch shim for quantic headless bundle (#3779)
[SFINT-5481](https://coveord.atlassian.net/browse/SFINT-5481)

This PR aims to fix the issue that was occurring in Salesforce when LWS
is disabled, this issue occurred because of the absence of the
`AbortController`.

In order to fix this, we are building a specific headless bundle made
only for the Quantic library.
In this bundle that will only be used inside Salesforce, we are
injecting some ponyfills that are providing the features that are
normally provided by the
[AbortController](https://developer.mozilla.org/en-US/docs/Web/API/AbortController
).

These ponyfills that are injected using ESbuild, will allow the headless
code to run properly by executing abortable fetch calls even when the
Abort Controller web API is not available inside Salesforce.



https://github.com/coveo/ui-kit/assets/86681870/06d6e2b5-116b-4071-8c4c-cf7304509ac7



[SFINT-5481]:
https://coveord.atlassian.net/browse/SFINT-5481?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Sylvie Allain <58052881+sallainCoveo@users.noreply.github.com>