fix(headless): analytics payload of quick view and smart snippet feedback fixed (#3797)
## [SFINT-5499](https://coveord.atlassian.net/browse/SFINT-5499)

In this PR, I fixed the following issues that has been identified in the
analytics payload of the following events sent with the new event
protocol:

- Quick view open: for the quickview, the `actionCause ` should be
`preview` instead of `open`.

- Smart snippet feedback submit. for the Smart Snippet Feedback submit,
`liked` should be set to `false` instead of `true`.

- Also, I updated the action cause `itemClick` to be in camel case in
order to be consistent with all the other events with the same type sent
from the Headless library.

[SFINT-5499]:
https://coveord.atlassian.net/browse/SFINT-5499?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Sylvie Allain <58052881+sallainCoveo@users.noreply.github.com>