refactor(atomic): first phase search box suggestions refactor (#3787)
First phase of the work for search box genericization effort for Atomic
for Commerce support.

**Note target branch**: I am targeting KIT-3036 instead of master. I
will be opening other PR related to KIT-3036; When everything is done
and merged in KIT-3036, I will open one last PR to merge it in master
(which will be the sum of those individual PR). The last "big" code
review to bring that in master should then be just a formality.

Vast majority of the complexity in the search box lies around suggestion
management. There some system in play that powers/handle query
suggestions:

* Handling performing suggestions request themselves to the APIs, which
is done at the web component level with Headless controller. This is not
something that is changed in any meaningful manner with this PR. For
`atomic-commerce-search box`, this would be a close copy of the code in
`atomic-search-box-suggestions` and the likes.

* System that support custom query suggestion, as documented here:
https://docs.coveo.com/en/atomic/latest/usage/atomic-custom-query-suggestions/
. This is not something that is changed in any meaningful manner in this
PR; The same base mechanism with events/suggestion interface stays the
same. Note that this is something implementer often need for commerce
(providing custom query suggestions/instant results), and that I
envision we will need first day support with Atomic for Commerce. We
just need to make sure the new commerce search box web component that
will be created support the system in place; It is now mostly handled by
the new `suggestion-manager`: What's left in `atomic-search-box` is
listening to the DOM event itself, but letting the manager do what's
necessary to handle the event.

* Handling the statefulness of what's the current active
suggestion/displayed suggestions VS end user actions (keyboard, mouse):
This is what this PR attempts to genericize with a new
`suggestion-manager` class. This new class would be re-usable by
`atomic-commerce-search-box` to manage keyboard and mouse actions.

* Handling accessibility (also known as "aria-live"), which is about
announcing to screen readers the interactive nature of query suggestions
(new suggestions appeared, which suggestion is currently active, what
will happen when a suggestion is selected). For this PR, I've deleted
the code around this from the search box and commented out the tests
that verified this part of the search box. I will work on this in my
subsequent PR, and create an accessiblity-manager class.

Net net is mostly:
* Delete `search-box-common`
* Delete duplicate code from `atomic-search-box` and
`atomic-insight-search-box`
* Deleted code above gets re-arranged/reworked into `suggestion-manager`
(this class is not a complete rewrite, mostly code scavenge)
* Genericize some types regarding custom query suggestions system to
make it re-usable for `atomic-commerce-search-box`

https://coveord.atlassian.net/browse/KIT-3036

---------

Co-authored-by: Frederic Beaudoin <fbeaudoin@coveo.com>