feat(headless-cart): allow both productid and SKU when updating a cart item with the cart controller (#3809)
[LENS-1761](https://coveord.atlassian.net/browse/LENS-1761)

It was noticed that the `productId` emitted with the cart events
(`ec.cartAction` & `ec.purchase`) in Barca sports was actually the sku
of the product. To guarantee the uniqueness of the cart item in the
state and that we send the right information to EP, we decided to add
the `sku` property to the `updateItem` action.

Demo of linking the solution to Barca sports


https://github.com/coveo/ui-kit/assets/58052881/524973b7-ab52-4072-9b38-ce955f96d174


[LENS-1761]:
https://coveord.atlassian.net/browse/LENS-1761?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ