fix: `enableQuerySyntax` is not passed to recent-query-list controller (#3818)
https://coveord.atlassian.net/browse/KIT-3096

Whenever a submit action (Enter key pressed) is being recorded on a
recent-query suggestion, the `prepareForSearchWithQuery` search action
is triggered with the appropriate `enableQuerySyntax` value.

However, when a mouse click action is recorded on a recent-query
suggestion, the `prepareForSearchWithQuery` search action called by the
headless-recent-queries-list controller does not specify the
`enableQuerySyntax` value.

---------

Co-authored-by: Frederic Beaudoin <fbeaudoin@coveo.com>