fix(atomic): separator should not be consider for ARIA (#3846)
Spec: https://w3c.github.io/aria/#list

## Issue

The important bit is this:
<img width="444" alt="image"
src="https://github.com/coveo/ui-kit/assets/12366410/1b6c4893-37cd-4034-ba89-e4afc0933268">

Essentially, anything but a role=listitem is forbidden under a
`role=list`. `ol` is defacto a list, always (afaik). So `role=separator`
is a nogo.

Example here:
https://louis-bompart.github.io/automatic-octo-spoon/index.html

## Solution

The obvious: remove the faulty role.

Editorial choice: I don't think the separator has any value whatsoever
for a screen reader, so I elected to use `aria-hidden` to make it
'disappear' for all intent and purposes from the standpoint of a screen
reader or accessibility tools.

## Tests:

This PR includes flakiness proofing on the accessibility test of this
component (the method could be reused on as needed).

### Wait for component to be hydrated

The flakiness of the test was reproducable on the client-side, and was
indeed a 'false green'. This was caused by axe core running _before_ the
component was hydrated, and on the component element nonetheless. So
yeah: cool cucumber, testing no DOM is ARIA compliant :slowclap:

This add a protection on that regard: Do wait for the component to be
hydrated before doing anything else.

The wait for hydration is used 'as needed' on the green path (i.e when
the component actually tries to render something)

### Properly reset between retries.
Fun stuff: If you are on foo.com/acme.html and navigate to
foo.com/acme.html#, you're not doing a DOM unload/load.
Essentially, it's the same stuff with cy.visit. I noted a pollution of
components in-between tests, which would skew the result, I reckon.
To fix that, I added a `cy.visit('/')` (last choice: `about:blank` tries
to goto `localhost:3333/about:blank` in the CI, and `chrome://dino`
redirect to `https://chrome://dino`, which is very disappointing.)