fix(quantic): answer generation glitch (#3843)
# Bug Fix

[SVCC-3357](https://coveord.atlassian.net/browse/SVCC-3357)

[Internal Coveo
Documentation](https://docs.google.com/document/d/1gX-Kr0yP_Mqog4CT8OYZAbu25YQbkaewjMJw-LewGK0)
about the issue and the proposed fix

## Description

### TLDR

In Quantic. Removal of the loading in the quantic generative answering
component will prevent the UI from glitching

### The problem

In Quantic
When the generative answering functionalities are enabled, but the model
is asked a question it cannot answer, the stream opens and closes very
quickly. This cause a jarring movement in the UI.

### The fix

In Quantic
Removing the loading, letting the component appears only when the answer
begins to stream will prevent the component from appearing and
disappearing uselessly.

[SVCC-3357]:
https://coveord.atlassian.net/browse/SVCC-3357?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

## Before Demo

[simplescreenrecorder-2024-04-23_12.50.28.webm](https://github.com/coveo/ui-kit/assets/45688129/830bd45b-e1a5-486c-8730-a7ec49ab9f07)

## After Demo

[simplescreenrecorder-2024-04-23_12.58.04.webm](https://github.com/coveo/ui-kit/assets/45688129/e1044390-1804-4c3b-b685-681655d121dc)

Co-authored-by: Louis Bompart <lbompart@coveo.com>