fix(atomic): answer generation component glitch (#3841)
# Bug Fix

[SVCC-3357](https://coveord.atlassian.net/browse/SVCC-3357)

[Internal Coveo
Documentation](https://docs.google.com/document/d/1gX-Kr0yP_Mqog4CT8OYZAbu25YQbkaewjMJw-LewGK0)
about the issue and the proposed fix

## Description

### TLDR

In Atomic. Removal of the loading in the quantic generative answering
component will prevent the UI from glitching

### The problem

In Atomic
When the generative answering functionalities are enabled, but the model
is asked a question it cannot answer, the stream opens and closes very
quickly. This cause a jarring movement in the UI.

### The fix

In Atomic
Removing the loading, letting the component appears only when the answer
begins to stream will prevent the component from appearing and
disappearing uselessly.

**Before**

[simplescreenrecorder-2024-04-22_16.30.24.webm](https://github.com/coveo/ui-kit/assets/45688129/c61d2075-8390-414d-ac7f-1a554a8a3d9d)

**After**

[simplescreenrecorder-2024-04-22_16.31.55.webm](https://github.com/coveo/ui-kit/assets/45688129/2cc162b7-5fca-4ffc-92ec-a2091e165cd7)


![image](https://github.com/coveo/ui-kit/assets/45688129/06768e36-9195-4e8d-a877-29d152cac5cd)




[SVCC-3357]:
https://coveord.atlassian.net/browse/SVCC-3357?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

Co-authored-by: Danny Gauthier <dgauthier@WKS-002169.corp.coveo.com>