fix(headless): generated-answer-state expanded should be false (#3848)
SFINT-5520

The reason why it should be false, is because otherwise we have to send
an analytics event on load of the component to “make it collapsed”.

For example, if the component is collapsible=true when loading the first
time, the answer should be collapsed if it’s too long.

If expanded starts true in the state, this means that in order to have
it collapsed, we need to trigger the action generatedAnswer.collapse
from the headless controller. This action will also send an analytics
event. Which means that on every load of the component, if
collapsible=true there would be a “fake” analytics event as if the user
had hit the “Show Less” button.