feat(headless/commerce): support continuous ranges on numeric facets (#3879)
https://coveord.atlassian.net/browse/CAPI-720

Initially when we implemented commerce numeric facets, ranges had to be
explicitly specified in the CMH configuration.

It is now also possible to configure numeric facets with automatic
discrete ranges, or with a single continuous range in the CMH.

Every facet response now includes a `domain` property indicating the
`min` and `max` range values that will return products when a set as a
selected range's `start` and `end` values respectively.

The `domain` also includes an `increment` property that indicates the
number of "steps" between each value. This will always be `0` if the
range is continuous, or if there is no regularity between the discrete
ranges (e.g., if the discrete ranges overlap or "jump" steps).

The `domain` property is not absolutely necessary for discrete ranges,
although an implementer could leverage it to allow end users to "add"
and select valid arbitrary numeric ranges in a facet (e.g., by entering
valid start and end values in an input).

The `domain` is absolutely necessary to implement continuous facets
(e.g., sliders), however.

In this pull request, I added support for the `domain` on the request
and response types (and did some refactoring too...), as well as on the
numeric facet controller state (I did not expose the `increment`
property in the numeric facet controller state, as I don't see an
immediate use for it, though). I also exposed a `setRanges` method on
the numeric facet controller.

Corresponding PR in Barca Sports:
https://github.com/coveo/barca-sports/pull/243