[Quantic] RGA Show More/Show Less buttons (#3811)
[SFINT-5428]

Until the scratch org expires, you can try it out here:


https://enterprise-enterprise-2467-dev-ed.scratch.my.site.com/examples/s/relevant-generative-answering-example#q=how%20to%20resolve%20netflix%20connection%20with%20tivo&tab=All
(Valid until the 25th)

Adding a new option to the QuanticGeneratedAnswer component:
`collapsible`.
This option is `false` by default. 

When turned on, this option will cap the size of the generated answer
text to a maximum of 250px.
A new control will be added to expand and/or collapse the answer that
was generated.

Some details on the behavior:

- While the answer is collapsed, the rephrase, feedback and citations
controls are not visible.
- When expanding the answer, the above controls appear.
- The state of expanded/collapsed will reset on each query, rephrase of
RGA as designed.
- If the answer is actually shorter than 250px nothing will change to
the current behavior of RGA.
- The rephrase, feedback, and citations components will all be visible.
  - No show more/less button should be visible.
- The new button should render left of the RGA disclaimer (so on a row
below the Rephrase and citations)
- The new button should work correctly when the footer is `multiline`
- The cap to the answer height should look "clean" with a fade out, etc.

See videos and screenshots below for examples.

1. The component and the new control should be responsive:


https://github.com/coveo/ui-kit/assets/2488155/653fc524-3916-4961-8fa5-61f9b1899f94

2. General function of the component:



https://github.com/coveo/ui-kit/assets/2488155/75cbb774-9da4-43c5-befd-3b41ef8caccb



[SFINT-5428]:
https://coveord.atlassian.net/browse/SFINT-5428?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

Also found a fix for : SFINT-5469