feat(service): add action on insight result action click (#3845)
[SVCINT-2951](https://coveord.atlassian.net/browse/SVCINT-2951)

### Proposition
As we talked about a while back it makes sense to add a prop to the
insight-result-action component and dispatch custom events. This way,
custom integrations like Genesys, ServiceNow and SAP will be able to
hook on those events and execute custom processing without the need for
integration-specific components.

I followed this doc: https://stenciljs.com/docs/events

Ex:


https://github.com/coveo/ui-kit/assets/7459000/86e9140c-9cab-443d-8603-cd7d2a5378cf




[SVCINT-2951]:
https://coveord.atlassian.net/browse/SVCINT-2951?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: GitHub Actions Bot <>
Co-authored-by: Louis Bompart <lbompart@coveo.com>