feat(headless/commerce): use interval from response to determine whether numeric facet is continuous range (#3889)
https://coveord.atlassian.net/browse/CAPI-720

We can't merge this right now, though: I tested it and the API currently
returns "continuous" for numeric facets with explicitly specified
discrete ranges. Until this is fixed, the current way to do the check
(i.e., verifying that the facet's increment is 0 and that it contains a
single value) is more reliable.

EDIT: I verified with @camarois and tested locally in Barca Sports. It
turns out that the correct interval value will be returned when it is
included in the facet request. Initially (i.e., when the facet is not in
the facets array of the request body), the response will also contain
the correct interval so that it can be set for subsequent requests. This
means this is ready to merge.