fix(atomic): use variable to set font-weight of strong and th elements in generated rich answers (#3926)
# [SVCC-3766](https://coveord.atlassian.net/browse/SVCC-3766)

In some cases, the bold text in rich generated answers could be too
light or too heavy compared to other bold elements in the search
interface.

This PR updates the styling of `strong` and `th` elements in generated
rich answers to rely on the `--atomic-font-bold` variable and maintain a
consistent look across the search interface.



[SVCC-3766]:
https://coveord.atlassian.net/browse/SVCC-3766?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ