fix(atomic): update initializeBindings utility function for commerce custom components (#3933)
Custom components were previously only possible with classic `Bindings`
(ie: headless search engine). We need to make it possible with
`CommerceBindings` (headless commerce engine)

Modify the exported function to instead expose the fact that it is
generic under the hood.
Make it so that it defaults on `Bindings` (search engine use case) for
backward compatibility.

https://coveord.atlassian.net/browse/KIT-3156

---------

Co-authored-by: Frederic Beaudoin <fbeaudoin@coveo.com>