fix(headless): improved condition that needs to be satisfied before handling a generated answer stream (#3934)
[SFINT-5516](https://coveord.atlassian.net/browse/SFINT-5516)


### The issue

In Salesforce, where the Abort Controller is not available, we currently
use a a ponyfill to mimic the behaviour of the the Abort Controller and
to create an alternative `abortable fetch`.

However, this workaround we created is not able to really abort and the
stop the communication completely.
I tested and confirmed that a simple fetch request can be stopped using
our ponyfill. However the server sent communication that we initiate
with the `fetch-event-source` library to get the generated answer is not
completely stopped, the server is still sending messages even after
aborting the request.

Therefor, the following corner case is occurring sometimes when
triggering a new search query while a generated answer is still being
streamed:



https://github.com/coveo/ui-kit/assets/86681870/17d1b9ba-d738-4ece-842c-29d0a19a21b2


### The solution:

To solve this, we strengthened the condition that needs to be satisfied
before taking a generated answer stream in consideration.

Now we first check if `abortController.signal.aborted` is set to `true`,
if it is the case we discard the stream even if its request is not
completely aborted.




https://github.com/coveo/ui-kit/assets/86681870/37cc4fc5-fb65-4f88-8a8d-d15d66e55c11





[SFINT-5516]:
https://coveord.atlassian.net/browse/SFINT-5516?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Sylvie Allain <58052881+sallainCoveo@users.noreply.github.com>
Co-authored-by: Louis Bompart <lbompart@coveo.com>