feat(atomic-react): add commerce components (#3944)
- Mimick what's there for other components but for commerce
- Little hack/patch in react-output-target to allow commerce components.
- Workaround to allow users 'bring-your-own-engine'. 
- San-check w/ sample. SSB behave oddly.
<img width="1728" alt="image"
src="https://github.com/coveo/ui-kit/assets/12366410/a91f1017-d43d-44d5-a78a-0cccf7ad624e">
It overlap the navbar.
This is due to the lack of textarea (which we can't put on because we
did not want to include textarea in the prop).
See #3923 for details.

True-fix will come with the layout.

---------

Co-authored-by: GitHub Actions Bot <>