feat(quantic): toggle RGA is now an option, false by default (#3943)
SFINT-5506

The toggle button on the Quantic Generated Answer component is now an
option on the component.
It is false by default, which means all existing Quantic RGA components
will see their Toggle button disappear. They can re-add it with an
option if they really need it.

This change was vetted by @operecoveo , since the usage is almost null,
we feel safe to hide the button by default.

The Toggle button WAS here 

![image](https://github.com/coveo/ui-kit/assets/2488155/35a2c391-52fa-49b8-a71b-7003ed0d4329)