fix(headless): send referrer as null when originLevel3 is default (#3974)
**Context**
The server event schema was recently tightened to require document
referrer to be a url if specified. The schema definition is available
[here](https://github.com/coveo-platform/analytics_schema/blob/master/schemas/common/types.json#L289).

Headless' behavior is to forward the originLevel3 value, which by
default is set to the string `default`.
This causes all server events sent by headless to be marked as invalid.

**Approach**
When the `analyticsMode` is `next` and originLevel3 is `default`,
headless will send `null` instead of `default`.