docs(headless): commerce controllers and actions (#3913)
https://coveord.atlassian.net/browse/KIT-3140

Discussed on slack mostly with @anthonydelage @fbeaudoincoveo and
@louis-bompart
We want to generate docs for the new headless commerce controllers, but
keep the documentation neither searchable nor accessible through the
TOC. Links will be shared with key customer/partners who have already
been shared a link to the GitHub repo.

This PR extracts docs from Headless in the ui-kit project, and this
[ticket](https://coveord.atlassian.net/browse/DOC-14905) will turn that
into docs in the doc_jekyll-public-site project.

Docs preview:
https://docsstaging.coveo.com/DOC-14905-demo-may-16/headless/latest/reference/commerce/

---------

Co-authored-by: Nicholas Labarre <Spuffynism@users.noreply.github.com>