chore(quantic):  fix daily quantic package creation task (#3949)
## [SFINT-5513](https://coveord.atlassian.net/browse/SFINT-5513)

The Quantic Package Creation task has been failing repeatedly lately:
https://github.com/coveo/ui-kit/actions/workflows/create-quantic-package.yml

The issue started appearing after starting to use the `Setup` action in
this workflow. The `Setup` actions uses the NX cache in order to gain
time and avoid installing and building everything every time.

However, I noticed that we are caching only the `node_modules`. and for
Quantic to be built correctly it needs the static resources to be
created out of the node_modules.


To fix that, I added a step to the `setup` action in order to also cache
the `static resources` folder.

Another solution could just be running `npx nx run quantic:build` in the
`Create Quantic Package` workflow.
Let me know which solution you prefer or if you have any better ideas.

[SFINT-5513]:
https://coveord.atlassian.net/browse/SFINT-5513?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Sylvie Allain <58052881+sallainCoveo@users.noreply.github.com>